f104_consumer/requirement.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
idna==2.9
itsdangerous==1.1.0
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
pandas==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
Werkzeug==1.0.1
html5lib
kafka-python
redis